Update string to enum
