+++ extensions/schema-compare/src/dialogs/schemaCompareDialog.ts
@@ -118,7 +118,7 @@ export class SchemaCompareDialog {
 				ownerUri: ownerUri,
 				projectFilePath: '',
 				targetScripts: [],
-				folderStructure: '',
+				folderStructure: mssql.ExtractTarget.none,
 				packageFilePath: '',
 				dataSchemaProvider: '',
 				connectionDetails: undefined,
@@ -133,7 +133,7 @@ export class SchemaCompareDialog {
 				ownerUri: '',
 				projectFilePath: '',
 				targetScripts: [],
-				folderStructure: '',
+				folderStructure: mssql.ExtractTarget.none,
 				dataSchemaProvider: '',
 				packageFilePath: this.sourceTextBox.value,
 				connectionDetails: undefined
@@ -144,7 +144,7 @@ export class SchemaCompareDialog {
 				projectFilePath: this.sourceTextBox.value,
 				targetScripts: await this.getProjectScriptFiles(this.sourceTextBox.value),
 				dataSchemaProvider: await this.getDatabaseSchemaProvider(this.sourceTextBox.value),
-				folderStructure: '',
+				folderStructure: mssql.ExtractTarget.none,
 				serverDisplayName: '',
 				serverName: '',
 				databaseName: '',
@@ -165,7 +165,7 @@ export class SchemaCompareDialog {
 				databaseName: this.targetDatabaseDropdown.value.toString(),
 				ownerUri: ownerUri,
 				projectFilePath: '',
-				folderStructure: '',
+				folderStructure: mssql.ExtractTarget.none,
 				targetScripts: [],
 				packageFilePath: '',
 				dataSchemaProvider: '',
@@ -180,7 +180,7 @@ export class SchemaCompareDialog {
 				databaseName: '',
 				ownerUri: '',
 				projectFilePath: '',
-				folderStructure: '',
+				folderStructure: mssql.ExtractTarget.none,
 				targetScripts: [],
 				dataSchemaProvider: '',
 				packageFilePath: this.targetTextBox.value,
@@ -190,7 +190,7 @@ export class SchemaCompareDialog {
 			this.schemaCompareMainWindow.targetEndpointInfo = {
 				endpointType: mssql.SchemaCompareEndpointType.Project,
 				projectFilePath: this.targetTextBox.value,
-				folderStructure: this.targetStructureDropdown!.value as string,
+				folderStructure: mapExtractTargetEnum(<string>this.targetStructureDropdown!.value),
 				targetScripts: await this.getProjectScriptFiles(this.targetTextBox.value),
 				dataSchemaProvider: await this.getDatabaseSchemaProvider(this.targetTextBox.value),
 				serverDisplayName: '',
@@ -1015,3 +1015,18 @@ function isNullOrUndefined(val: any): boolean {
 	return val === null || val === undefined;
 }
 
+/**
+ * Function to map folder structure string to enum
+ * @param inputTarget folder structure in string
+ * @returns folder structure in enum format
+ */
+export function mapExtractTargetEnum(inputTarget: string): mssql.ExtractTarget {
+	switch (inputTarget) {
+		case loc.file: return mssql.ExtractTarget.file;
+		case loc.flat: return mssql.ExtractTarget.flat;
+		case loc.objectType: return mssql.ExtractTarget.objectType;
+		case loc.schema: return mssql.ExtractTarget.schema;
+		case loc.schemaObjectType: return mssql.ExtractTarget.schemaObjectType;
+		default: return mssql.ExtractTarget.none;
+	}
+}
